忘记提交了，目前已实现全部category业务以及article的查询列表，新增，查询详情功能，正在实现编辑功能

+++ TouHouBlog-Backend/src/main/java/BlogBack/service/CategoryService.java
@@ -24,7 +24,7 @@ public interface CategoryService {
      * @param categoryDTO
      * @param id
      */
-    void update(CategoryDTO categoryDTO, Integer id);
+    void update(CategoryDTO categoryDTO, Long id);
 
     /**
      * 根据id查询分类信息
